fix: underflow math error on harvest

+++ interfaces/Curve/ICurve.sol
@@ -23,15 +23,15 @@ interface ICurve {
     ) external returns (uint256);
 
     function exchange(
-        int128 i,
-        int128 j,
+        uint256 i,
+        uint256 j,
         uint256 dx,
         uint256 min_dy
     ) external returns (uint256);
 
     function get_dy(
-        int128 i,
-        int128 j,
+        uint256 i,
+        uint256 j,
         uint256 dx
     ) external returns (uint256);
 }
